9. Authentication and Authorization - 22- Exercise - En moviesTable.jsx en el metodo render() se hizo una copia del array columns (utilizando slice) y le agregamos el boton de delete si el usuario esta logueado y si esta logueado como admin, el array movies que viene de movies.jsx solo se manda a llamar una vez al servidor, por lo que hay que tener en cuenta eso, se hizo una copia de columns en moviesTable.jsx porque si no se hiciera esto cada vez que le dieramos delete un nuevo obj de delete button se agregaria, cosa que no queremos, la solucion de mosh es un poco diferente y evita que tengamos que copiar la variable columns

+++ src/components/moviesTable.jsx
@@ -2,6 +2,7 @@ import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import Table from "./common/table";
 import Like from "./common/like";
+import auth from '../services/authService';
 
 class MoviesTable extends Component {
     columns = [
@@ -18,26 +19,33 @@ class MoviesTable extends Component {
             content: movie => (
                 <Like liked={movie.liked} onClick={() => this.props.onLike(movie)} />
             )
-        },
-        {
-            key: "delete",
-            content: movie => (
-                <button
-                    onClick={() => this.props.onDelete(movie)}
-                    className="btn btn-danger btn-sm"
-                >
-                    Delete
-                </button>
-            )
         }
     ];
 
     render() {
         const { movies, onSort, sortColumn } = this.props;
 
+        let columns = this.columns.slice()
+        // console.log(columns)
+        const user = auth.getCurrentUser()
+        // console.log(user)
+        if (user && user.isAdmin) {
+            columns.push({
+                key: "delete",
+                content: movie => (
+                    <button
+                        onClick={() => this.props.onDelete(movie)}
+                        className="btn btn-danger btn-sm"
+                    >
+                        Delete
+                    </button>
+                )
+            })
+        }
+
         return (
             <Table
-                columns={this.columns}
+                columns={columns}
                 data={movies}
                 sortColumn={sortColumn}
                 onSort={onSort}
